move section structure to a separate module
